Created an abstraction layer to handle movement concerns
Creates better separation of concerns so game elements can declaratively specify movement and the movement engine will convert it into actual movement actions. This has a particularly big impact in the unit tests for basic and speedy wheels, where we can now just drop in a mock mover and assert it was called the correct number of times.